backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
mysqlclient==2.2.7
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0